게시글 삭제불가 버그 수정 & 댓글 허용 여부 체크 기능 추가
- 게시글 작성, 수정시 댓글 허용 여부를 함께 전송하도록 합니다.
- 게시글 수정 시 기존의 최소 조회 권한과 댓글 허용 여부에 대한 체크 표시가 없던 문제를 수정합니다.
- 아무런 첨부파일을 올리지 않은 경우 게시글 제거가 불가한 문제를 수정합니다.
- 게시글 신규 작성 화면에 취소 버튼을 추가합니다.

+++ src/main/java/kr/pe/hayarobys/nox/common/upload/UploadServiceImpl.java
@@ -106,7 +106,7 @@ public class UploadServiceImpl implements UploadService{
 	public void deleteFileGroupByFileGroupNoList(List<Integer> fileGroupNoList){		
 		// 파일 목록 조회
 		List<FileVO> fileList = uploadDAO.selectFileByFileGroupNoList(fileGroupNoList);
-
+		
 		// 파일 목록들의 모든 물리 파일 제거
 		List<Integer> fileNoList = new ArrayList<Integer>();
 		for(FileVO fileVO : fileList){
@@ -117,7 +117,9 @@ public class UploadServiceImpl implements UploadService{
 		}
 		
 		// DB에서 파일 목록 제거
-		uploadDAO.deleteFileByFileNoList(fileNoList);
+		if(fileList.size() > 0){
+			uploadDAO.deleteFileByFileNoList(fileNoList);
+		}
 		
 		// DB에서 파일 그룹 목록 제거
 		uploadDAO.deleteFileGroupByFileGroupNoList(fileGroupNoList);
